requirements.txt
httpx>=0.27.0
pyyaml>=6.0
pandas>=2.2.0
python-dotenv>=1.0.0
